Fixing Function Signatures
VQE.py -> Fixed TypeError when importing VQE. The object `list` is
a class and they are of type type. You need to use typing.List for type
hints instead

utils.py -> If you dont provide molecules_qubitop.npy in the repo, you
shouldn't have the default behavior be to read that file. molecules'
function signature has been modified to set the default option of load
to be False.